+++ packaging/suse/portusctl/lib/cli.rb
@@ -1,0 +1,93 @@
+# Class implementing the cli interface of portusctl
+class Cli < Thor
+  desc "setup", "Configure Portus"
+  option :secure, type: :boolean, default: true,
+    desc: "Toggle SSL usage for Portus"
+  # SSL certificate options
+  option "ssl-organization",
+    desc:    "SSL certificate: organization",
+    default: "SUSE Linux GmbH" # gensslcert -o
+  option "ssl-organization-unit",
+    desc:    "SSL certificate: organizational unit",
+    default: "SUSE Portus example" # gensslcert -u
+  option "ssl-email",
+    desc:    "SSL certificate: email address of webmaster",
+    default: "kontact-de@novell.com" # gensslcert -e
+  option "ssl-country",
+    desc:    "SSL certificate: country (two letters)",
+    default: "DE" # gensslcert -c
+  option "ssl-city",
+    desc:    "SSL certificate: city",
+    default: "Nuremberg" # gensslcert -l
+  option "ssl-state",
+    desc:    "SSL certificate: state",
+    default: "Bayern" # gensslcert -s
+  # DB options
+  option "db-host", desc: "Database: host", default: "localhost"
+  option "db-username", desc: "Database: username", default: "portus"
+  option "db-password", desc: "Database: password", default: "portus"
+  option "db-name", desc: "Database: name", default: "portus_production"
+  # Registry
+  option "local-registry", desc: "Configure Docker registry running locally",
+    type: :boolean, default: false
+  # LDAP
+  option "ldap-enable", desc: "LDAP: enable", type: :boolean, default: false
+  option "ldap-hostname", desc: "LDAP: server hostname"
+  option "ldap-port", desc: "LDAP: server port", default: "389"
+  option "ldap-base",
+    desc:    "LDAP: base",
+    default: "ou=users, dc=example, dc=com"
+  option "ldap-guess-email",
+    desc:    "LDAP: guess email address",
+    type:    :boolean,
+    default: false
+  option "ldap-guess-email-attr",
+    desc: "LDAP: attribute to use when guessing email address"
+  # MAILER
+  option "email-from",
+    desc:    "MAIL: sender address",
+    default: "portus@#{HOSTNAME}"
+  option "email-name", desc: "MAIL: sender name", default: "Portus"
+  option "email-reply-to",
+    desc:    "MAIL: reply to address",
+    default: "no-reply@#{HOSTNAME}"
+  # GRAVATAR
+  option "gravatar", desc: "Enable Gravatar usage", type: :boolean, default: true
+  # FIRST USER
+  option "first-user-admin",
+    desc:    "Make the first registered user an admin",
+    type:    :boolean,
+    default: true
+
+  def setup
+    if Process.uid != 0
+      warn "Must run as root user"
+      exit 1
+    end
+
+    if options["ldap-enable"] && (options["ldap-hostname"].nil? || options["ldap-hostname"].empty?)
+      warn "LDAP support is enabled but you didn't specify a value for ldap-hostname"
+      exit 1
+    end
+
+    configure = Configurator.new(options)
+    configure.apache
+    configure.ssl
+    configure.database_config
+    configure.registry if options["local-registry"]
+    configure.secrets
+    configure.config_local
+    configure.create_database
+    configure.services
+  end
+
+  desc "rake ARGS...", "Run a rake task against Portus"
+  def rake(*args)
+    if args.empty?
+      warn "You mush provide at least an argument"
+      exit 1
+    end
+
+    Runner.bundler_exec("rake", args, {})
+  end
+end