+++ packaging/suse/scripts/configure_crono.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-systemctl enable portus_crono
-systemctl start portus_crono
-