+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Upcoming Version
 
+- SUSE RPM: provide `portusctl` tool to simplify the initial setup of Portus
 - Portus will now lock users' accounts that have failed too many times on
 login. See PR [#330](https://github.com/SUSE/Portus/pull/330).
 - Added a mechanism of password recovery in case users forget about their